Adding config/ and mapping/ files for all satwind types (#15)
Added high-level encoders and mapping YAMLs to mapping/ and script and
bufr backend YAMLs to config/ for all remaining satwind types: AHI,
AVHRR, LEOGEO, MODIS, SEVIRI, VIIRS. These were all tested on both
`2021080100` and `2023080100` dump files to confirm that proper
satellites are extracted from the BUFR file at their respective time in
their lifecycle (e.g., SEVIRI winds are collected for M8 and M11 in 2021
and for M9 and M10 in 2023). Both `script_backend` and `bufr_backend`
tests were performed. The only untested option is NOAA-21/VIIRS, which
wasn't available for either dataset.

Most of these were tested on `nproc=1`, due to the emergence of MPI i/o
errors in writing the output netCDF files that are appearing for some
satellites. These errors are expected to be related to issues upstream
of and unrelated to the SPOC files.

---------

Co-authored-by: Emily Liu <emily.liu@noaa.gov>