PYTest Integration (#56)
This pull request adds PYTests to the SPOC project. The following files
are added:

tests/conftest.py
tests/test_functions.py
tests/test_cases.py             *Add your test cases here

The following tool was added to create a minimized version of the BUFR
file. It keeps a certain number of (default 1000) category/subset data
observations. It's a little smarter in that it keeps data for all the
categories and subsets in that category (optional) in order to preserve
the variety of the data in the BUFR file:

tools/bufr_minimizer/bufr_minimizer.py

---------

Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>