Feature/obs builder prepbufr (#29)
This PR updates the prepbufr_obs_builder.py

The main change is the object ```_replace_timestamp``` overhaul. I
needed a way to deal with missing dateTime values, which were not dealt
with originally. I replaced values where dateTime is missing with
np.int64(0). While ADPSFC and SFCSHP did not need this, AIRCFT does.
Please note AIRCFT is not in this PR, as the group_by issues are still
prevalent. When the group_by issue is fixed, I will add it in a separate
PR.

This fixes issue #27 and #28 and #5

---------

Co-authored-by: Nicholas Esposito <esposito@hercules-login-2.hpc.msstate.edu>
Co-authored-by: Nicholas Esposito <esposito@hercules-login-3.hpc.msstate.edu>
Co-authored-by: emilyhcliu <emily.liu@noaa.gov>
Co-authored-by: Nicholas Esposito <esposito@hercules-login-4.hpc.msstate.edu>