Feature/obs builder adpupa prepbufr (#40)
This PR adds following four files for the ADPUPA prepBUFR:

- ~/dump/mapping/prepbufr_adpupa.py
- ~/dump/mapping/prepbufr_adpupa.yaml
- ~/ush/test/config/bufr_bufr4backend_adpupa_prepbufr.yaml
- ~/ush/test/config/bufr_script4backend_adpupa_prepbufr.yaml

All four jobs (bufr2netcdf, bufr4backend, script2netcdf, and
script4backend) were run and tested using serial execution. When using
parallel execution for bufr2netcdf, the number of observations (and the
output file size) decreases as the number of nodes increases:

- gdas.t00z.adpupa.tm00.nc         ---->  17760343
- gdas.t00z.adpupa.tm00_2.nc      ---->  17244289
- gdas.t00z.adpupa.tm00_4.nc      ---->  16549247 
- gdas.t00z.adpupa.tm00_6.nc      ---->  16349380
- gdas.t00z.adpupa.tm00_8.nc      ---->  15913611 

The dateTime variable is computed both in the YAML configuration and in
the Python script. In the Python script, it is calculated using the
compute_datetime function imported from prepbufr_obs_builder.py, and the
final value is inserted into the output container. Currently, the
function uses "0" to represent missing values.

Validation tests for this data type have already been completed.

The spoc_python_coding_norms test was run to verify compliance with
Pycodestyle standards for the Python script(s). The results were as
follows:

The following tests passed:
        _spoc_python_coding_norms_

**100% tests passed, 0 tests failed out of 1**

---------

Co-authored-by: Praveen Singh <pkumar@hercules-login-1.hpc.msstate.edu>
Co-authored-by: PraveenKumar-NOAA <Praveen.Kumar@noaa.gov>
Co-authored-by: emilyhcliu <emily.liu@noaa.gov>